requirements.txt
discord.py
llama-cpp-python
llama-cpp-python[server]
openai
python-dotenv
tinydb